refactor: split tool_dispatch.rs into three focused modules (#643)
Extract sub-agent invocation and approval flow into dedicated files:

- sub_agent_dispatch.rs (410 lines): execute_sub_agent, run_bg_agent
  Handles InvokeAgent execution, background spawning, worktree
  provisioning, sub-agent caching, and the inference loop for
  sub-agent tool calls.

- approval_flow.rs (90 lines): request_approval, handle_ask_user
  Async request/response for tool approvals and AskUser interaction.

- tool_dispatch.rs (777 lines, was 1243): sequential, parallel,
  split-batch execution, result recording, file lifecycle tracking.

No behavior change. All 286 tests pass.

Dependency graph:
  tool_dispatch → sub_agent_dispatch → approval_flow
               → approval_flow
